src/requirements.txt
cloudshell-automation-api>=8.3,<8.4
cloudshell-core>=2.2,<2.3
cloudshell-shell-core>=4.0,<4.1
cloudshell-networking-devices>=2.5,<2.6
